Fix issues on object arrays with numpy 1.20+
np.asarray(..., dtype=np.object) with numpy 1.20+ behaves differently
than with previous verisons of numpy when elements are array-like.
Starting from numpy 1.20+, copy will occur when the elements of a list
is an array-like (e.g., cuda tensor or numpy array).

As a result, with numpy 1.20+ an unexpected shape error can happen.

Reference:

https://numpy.org/devdocs/release/1.20.0-notes.html#array-coercion-restructure